fix: 자료 한국어화
